Attempted to phase out SystemView, using Sampler directly
